// apps/web/src/shared/api/generated/model/deliveryControllerStatisticsParams.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * book-nest API
 * REST API for the book-nest project
 * OpenAPI spec version: 1.0
 */
import type { DeliveryControllerStatisticsCurrency } from "./deliveryControllerStatisticsCurrency";
import type { DeliveryControllerStatisticsIncludeCancelled } from "./deliveryControllerStatisticsIncludeCancelled";
import type { DeliveryControllerStatisticsStatus } from "./deliveryControllerStatisticsStatus";

export type DeliveryControllerStatisticsParams = {
  currency?: DeliveryControllerStatisticsCurrency;
  /**
   * @pattern ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))$
   */
  from?: string;
  includeCancelled?: DeliveryControllerStatisticsIncludeCancelled;
  status?: DeliveryControllerStatisticsStatus;
  /**
   * @maxLength 200
   */
  store?: string;
  /**
   * @pattern ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))$
   */
  to?: string;
};